requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
dspy-ai>=2.3.0
llama-index>=0.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
numpy>=1.26.4
scikit-learn>=1.4.1.post1
statsmodels>=0.14.1